# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Accounting Platform API
  version: 0.1.0
  description: >
    REST API for the AbhashTech multi-tenant accounting and bookkeeping
    platform with double-entry ledger, invoicing, expenses, payroll, tax,
    reporting, imports, backups, and investment accounting.
  contact:
    name: AbhashTech Accounting
    url: https://accounting.abhashtech.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.accounting.abhashtech.com/api/v1
    description: Production placeholder
...
security:
  - bearerAuth: []
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: List organizations available to the current user
      operationId: listOrganizations
      responses:
        '200':
          description: Organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Organizations]
      summary: Create an organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Created organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        default:
          $ref: '#/components/responses/Error'
...
  /organizations/{organizationId}/invoices/{invoiceId}/payments:
    get:
      tags: [Invoices]
      summary: List customer payments applied to an invoice
      operationId: listCustomerPayments
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerPayment'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Invoices, Ledger]
      summary: Record a customer payment against an invoice
      operationId: recordCustomerPayment
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordPaymentRequest'
      responses:
        '201':
          description: Recorded customer payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPayment'
        default:
          $ref: '#/components/responses/Error'
...
  /organizations/{organizationId}/bills/{billId}/payments:
    get:
      tags: [Expenses]
      summary: List vendor payments applied to a bill
      operationId: listVendorPayments
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: billId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vendor payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VendorPayment'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Expenses, Ledger]
      summary: Record a vendor payment against a bill
      operationId: recordVendorPayment
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: billId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordPaymentRequest'
      responses:
        '201':
          description: Recorded vendor payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorPayment'
        default:
          $ref: '#/components/responses/Error'
...
  /organizations/{organizationId}/investments/average-cost-sales:
    post:
      tags: [Investments]
      summary: Record an average-cost pooled investment sale
      operationId: sellAverageCost
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellAverageCostRequest'
      responses:
        '201':
          description: Recorded average-cost sale
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AverageCostSaleResult'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
              example: Request validation failed
            details:
              type: array
              items:
                type: object
                additionalProperties: true
...
    Organization:
      type: object
      required: [id, name, base_currency, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        base_currency:
          type: string
          example: USD
        created_at:
          type: string
          format: date-time
    CreateOrganizationRequest:
      type: object
      required: [name, base_currency]
      properties:
        name:
          type: string
        base_currency:
          type: string
          minLength: 3
          maxLength: 3
...
    CustomerPayment:
      type: object
      required: [id, organization_id, invoice_id, payment_number, payment_date, currency, amount_minor, payment_account_id, journal_transaction_id]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        invoice_id:
          type: string
          format: uuid
        payment_number:
          type: string
          example: RCPT-001
        payment_date:
          type: string
          format: date-time
        payment_method:
          type: string
          example: bank_transfer
        reference:
          type: string
          example: UTR123456
        currency:
          type: string
          example: INR
        amount_minor:
          type: integer
          format: int64
          example: 11800
        payment_account_id:
          type: string
          format: uuid
        journal_transaction_id:
          type: string
          format: uuid
    RecordPaymentRequest:
      type: object
      required: [payment_number, payment_date, amount_minor, payment_account_id]
      properties:
        payment_number:
          type: string
          example: RCPT-001
        payment_date:
          type: string
          format: date
        payment_method:
          type: string
          example: bank_transfer
        reference:
          type: string
          example: UTR123456
        currency:
          type: string
          example: INR
        amount_minor:
          type: integer
          format: int64
          minimum: 1
          example: 11800
        payment_account_id:
          type: string
          format: uuid
...
    VendorPayment:
      type: object
      required: [id, organization_id, bill_id, payment_number, payment_date, currency, amount_minor, payment_account_id, journal_transaction_id]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        bill_id:
          type: string
          format: uuid
        payment_number:
          type: string
          example: VPAY-001
        payment_date:
          type: string
          format: date-time
        payment_method:
          type: string
          example: bank_transfer
        reference:
          type: string
          example: UTR987654
        currency:
          type: string
          example: INR
        amount_minor:
          type: integer
          format: int64
          example: 11800
        payment_account_id:
          type: string
          format: uuid
        journal_transaction_id:
          type: string
          format: uuid
...
    InvestmentDisposition:
      type: object
      required: [id, organization_id, investment_lot_id, sale_date, quantity_millis, proceeds_minor, allocated_cost_basis_minor, realized_gain_loss_minor, currency]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        investment_lot_id:
          type: string
          format: uuid
        sale_date:
          type: string
          format: date-time
        quantity_millis:
          type: integer
          format: int64
        proceeds_minor:
          type: integer
          format: int64
        allocated_cost_basis_minor:
          type: integer
          format: int64
        realized_gain_loss_minor:
          type: integer
          format: int64
        currency:
          type: string
          example: INR
        journal_transaction_id:
          type: string
          format: uuid
          nullable: true
        notes:
          type: string
...
    SellAverageCostRequest:
      type: object
      required: [account_id, symbol, sale_date, quantity_millis, proceeds_minor]
      properties:
        account_id:
          type: string
          format: uuid
        symbol:
          type: string
          example: LIQUIDFUND
        currency:
          type: string
          example: INR
        sale_date:
          type: string
          format: date
        quantity_millis:
          type: integer
          format: int64
        proceeds_minor:
          type: integer
          format: int64
        proceeds_account_id:
          type: string
          format: uuid
          description: Optional account debited for proceeds when posting the sale to GL.
        gain_loss_account_id:
          type: string
          format: uuid
          description: Optional income/expense account used for realized gain or loss when posting the sale to GL.
        notes:
          type: string
    AverageCostSaleResult:
      type: object
      required: [dispositions, quantity_millis, proceeds_minor, allocated_cost_basis_minor, realized_gain_loss_minor]
      properties:
        dispositions:
          type: array
          items:
            $ref: '#/components/schemas/InvestmentDisposition'
        quantity_millis:
          type: integer
          format: int64
        proceeds_minor:
          type: integer
          format: int64
        allocated_cost_basis_minor:
          type: integer
          format: int64
        realized_gain_loss_minor:
          type: integer
          format: int64
        journal_transaction_id:
          type: string
          format: uuid
          nullable: true
...